feat(ExerciseQuestion): add accept and reject functionality for exercise questions

+++ src/main/java/vn/fu_ohayo/repository/ExerciseResultRepository.java
@@ -8,7 +8,7 @@ import vn.fu_ohayo.entity.ExerciseResult;
 import java.util.List;
 
 @Repository
-public interface ExerciseResultRepository extends JpaRepository<ExerciseResult, Long> {
+public interface ExerciseResultRepository extends JpaRepository<ExerciseResult, Integer> {
     List<ExerciseResult> findAllByUserAndLessonExerciseIn(User user, List<LessonExercise> lessonExercises);
      int countByUserAndLessonExerciseIn(User user, List<LessonExercise> lessonExercises);
     @Query(value = "SELECT er FROM ExerciseResult er " +
